+++ test/unit/lib/runner/writers/index-tests.js
@@ -7,6 +7,7 @@ const expect = require('chai').expect;
 describe('lib/runner/writers/index.js', function() {
   let json;
   let actionJson;
+  let jUnit;
 
   beforeEach(function() {
     mockery.enable({useCleanCache: true});
@@ -14,9 +15,11 @@ describe('lib/runner/writers/index.js', function() {
 
     json = sinon.stub();
     actionJson = sinon.stub();
+    jUnit = sinon.stub();
 
     mockery.registerMock('./write-json-to-disk.js', json);
     mockery.registerMock('./action-json-writer.js', actionJson);
+    mockery.registerMock('./j-unit-writer.js', jUnit);
   });
 
   afterEach(function() {
@@ -25,10 +28,10 @@ describe('lib/runner/writers/index.js', function() {
     mockery.disable();
   });
 
-  it('should export 2 items on an object', function() {
+  it('should export 3 items on an object', function() {
     let result = require('../../../../../lib/runner/writers');
 
-    expect(Object.getOwnPropertyNames(result).length).to.equal(2);
+    expect(Object.getOwnPropertyNames(result).length).to.equal(3);
   });
 
   it('should have the property \'JSON\' with the value from requiring'
@@ -43,5 +46,12 @@ describe('lib/runner/writers/index.js', function() {
     let result = require('../../../../../lib/runner/writers');
 
     expect(result.actionJSON).to.deep.equal(actionJson.write);
+  });
+
+  it('should have the property \'JUnit\' with the value from requiring'
+    + ' \'./j-unit-writer.js\'', function() {
+    let result = require('../../../../../lib/runner/writers');
+
+    expect(result.JUnit).to.deep.equal(jUnit.write);
   });
 });